Use u_int32_t for token count values.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk/bogofilter@3446 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ src/msgcounts.h
@@ -22,7 +22,7 @@ extern	yylex_t	 read_msg_count_line;
 extern	bool	 msgcount_more(void);
 
 extern	char	*msg_count_text;
-extern	int	 msg_count_leng;
+extern	size_t	 msg_count_leng;
 
 /* Function prototypes */
